requirements.txt
allure-pytest==2.15.3
pytest==9.0.2
selenium==4.41.0
webdriver-manager==4.0.2